modify: separate fetchuser apart from question api
