+++ app/assets/javascripts/webpack/admin/questions/index.js
@@ -1112,7 +1112,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _axios = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _apiBase = __webpack_require__(/*! ./apiBase */ \"./src/api/apiBase.js\");\n\nvar _apiBase2 = _interopRequireDefault(_apiBase);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar QuestionApi = function (_ApiBase) {\n  _inherits(QuestionApi, _ApiBase);\n\n  function QuestionApi() {\n    _classCallCheck(this, QuestionApi);\n\n    return _possibleConstructorReturn(this, (QuestionApi.__proto__ || Object.getPrototypeOf(QuestionApi)).apply(this, arguments));\n  }\n\n  _createClass(QuestionApi, null, [{\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/' + ('' + questionId)).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchUser',\n    value: function fetchUser() {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/users/entry').then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchAnswer',\n    value: function fetchAnswer(questionId, answerId, userId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        var token = document.getElementsByName(\"csrf-token\")[0].getAttribute(\"content\");\n        _axios2.default.defaults.headers.common[\"X-CSRF-Token\"] = token;\n        _axios2.default.put(baseUrl + '/questions' + ('/' + questionId), {\n          question: {\n            user_id: userId,\n            question_id: questionId,\n            answer: answerId\n          }\n        }).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }]);\n\n  return QuestionApi;\n}(_apiBase2.default);\n\nexports.default = QuestionApi;\n\n//# sourceURL=webpack:///./src/api/question.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _axios = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _apiBase = __webpack_require__(/*! ./apiBase */ \"./src/api/apiBase.js\");\n\nvar _apiBase2 = _interopRequireDefault(_apiBase);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar QuestionApi = function (_ApiBase) {\n  _inherits(QuestionApi, _ApiBase);\n\n  function QuestionApi() {\n    _classCallCheck(this, QuestionApi);\n\n    return _possibleConstructorReturn(this, (QuestionApi.__proto__ || Object.getPrototypeOf(QuestionApi)).apply(this, arguments));\n  }\n\n  _createClass(QuestionApi, null, [{\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/' + ('' + questionId)).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchAnswer',\n    value: function fetchAnswer(questionId, answerId, userId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        var token = document.getElementsByName(\"csrf-token\")[0].getAttribute(\"content\");\n        _axios2.default.defaults.headers.common[\"X-CSRF-Token\"] = token;\n        _axios2.default.put(baseUrl + '/questions' + ('/' + questionId), {\n          question: {\n            user_id: userId,\n            question_id: questionId,\n            answer: answerId\n          }\n        }).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }]);\n\n  return QuestionApi;\n}(_apiBase2.default);\n\nexports.default = QuestionApi;\n\n//# sourceURL=webpack:///./src/api/question.js?");
 
 /***/ }),
