+++ app/assets/javascripts/webpack/index.js
@@ -1123,7 +1123,19 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n    value: true\n});
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _axios = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _apiBase = __webpack_require__(/*! ./apiBase */ \"./src/api/apiBase.js\");\n\nvar _apiBase2 = _interopRequireDefault(_apiBase);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar QuestionApi = function (_ApiBase) {\n  _inherits(QuestionApi, _ApiBase);\n\n  function QuestionApi() {\n    _classCallCheck(this, QuestionApi);\n\n    return _possibleConstructorReturn(this, (QuestionApi.__proto__ || Object.getPrototypeOf(QuestionApi)).apply(this, arguments));\n  }\n\n  _createClass(QuestionApi, null, [{\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/' + ('' + questionId)).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchUser',\n    value: function fetchUser() {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/users/entry').then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchAnswer',\n    value: function fetchAnswer(questionId, answerId, userId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        var token = document.getElementsByName(\"csrf-token\")[0].getAttribute(\"content\");\n        _axios2.default.defaults.headers.common[\"X-CSRF-Token\"] = token;\n        _axios2.default.put(baseUrl + '/questions' + ('/' + questionId), {\n          question: {\n            user_id: userId,\n            question_id: questionId,\n            answer: answerId\n          }\n        }).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }]);\n\n  return QuestionApi;\n}(_apiBase2.default);\n\nexports.default = QuestionApi;\n\n//# sourceURL=webpack:///./src/api/question.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _axios = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _apiBase = __webpack_require__(/*! ./apiBase */ \"./src/api/apiBase.js\");\n\nvar _apiBase2 = _interopRequireDefault(_apiBase);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar QuestionApi = function (_ApiBase) {\n  _inherits(QuestionApi, _ApiBase);\n\n  function QuestionApi() {\n    _classCallCheck(this, QuestionApi);\n\n    return _possibleConstructorReturn(this, (QuestionApi.__proto__ || Object.getPrototypeOf(QuestionApi)).apply(this, arguments));\n  }\n\n  _createClass(QuestionApi, null, [{\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/' + ('' + questionId)).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }, {\n    key: 'fetchAnswer',\n    value: function fetchAnswer(questionId, answerId, userId) {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        var token = document.getElementsByName(\"csrf-token\")[0].getAttribute(\"content\");\n        _axios2.default.defaults.headers.common[\"X-CSRF-Token\"] = token;\n        _axios2.default.put(baseUrl + '/questions' + ('/' + questionId), {\n          question: {\n            user_id: userId,\n            question_id: questionId,\n            answer: answerId\n          }\n        }).then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }]);\n\n  return QuestionApi;\n}(_apiBase2.default);\n\nexports.default = QuestionApi;\n\n//# sourceURL=webpack:///./src/api/question.js?");
+
+/***/ }),
+
+/***/ "./src/api/user.js":
+/*!*************************!*\
+  !*** ./src/api/user.js ***!
+  \*************************/
+/*! no static exports found */
+/***/ (function(module, exports, __webpack_require__) {
+
+"use strict";
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _axios = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n\nvar _axios2 = _interopRequireDefault(_axios);\n\nvar _apiBase = __webpack_require__(/*! ./apiBase */ \"./src/api/apiBase.js\");\n\nvar _apiBase2 = _interopRequireDefault(_apiBase);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar UserApi = function (_ApiBase) {\n  _inherits(UserApi, _ApiBase);\n\n  function UserApi() {\n    _classCallCheck(this, UserApi);\n\n    return _possibleConstructorReturn(this, (UserApi.__proto__ || Object.getPrototypeOf(UserApi)).apply(this, arguments));\n  }\n\n  _createClass(UserApi, null, [{\n    key: 'fetchUser',\n    value: function fetchUser() {\n      var baseUrl = this.defaultBaseUrl();\n\n      return new Promise(function (resolve, reject) {\n        _axios2.default.get(baseUrl + '/questions/users/entry').then(function (response) {\n          return resolve(response.data);\n        }).catch(function (error) {\n          return reject(error.response.data);\n        });\n      });\n    }\n  }]);\n\n  return UserApi;\n}(_apiBase2.default);\n\nexports.default = UserApi;\n\n//# sourceURL=webpack:///./src/api/user.js?");
 
 /***/ }),
 
@@ -1135,7 +1147,7 @@ eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n
 /***/ (function(module, exports, __webpack_require__) {
 
 "use strict";
-eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"]) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError(\"Invalid attempt to destructure non-iterable instance\"); } }; }();\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _socket = __webpack_require__(/*! socket.io-client */ \"./node_modules/socket.io-client/lib/index.js\");\n\nvar _socket2 = _interopRequireDefault(_socket);\n\nvar _Question = __webpack_require__(/*! ./Question */ \"./src/components/Question.js\");\n\nvar _Question2 = _interopRequireDefault(_Question);\n\nvar _question = __webpack_require__(/*! ../api/question */ \"./src/api/question.js\");\n\nvar _question2 = _interopRequireDefault(_question);\n\nvar _jsCookie = __webpack_require__(/*! js-cookie */ \"./node_modules/js-cookie/src/js.cookie.js\");\n\nvar _jsCookie2 = _interopRequireDefault(_jsCookie);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar socket = _socket2.default.connect('http://localhost:3005');\n\n// Cookies.remove('user_name');\n// Cookies.remove('question_number');\nvar user_name = _jsCookie2.default.get('user_name') ? _jsCookie2.default.get('user_name') : 'guest';\nvar user_id = _jsCookie2.default.get('user_id') ? _jsCookie2.default.get('user_id') : 0;\nvar question_number = _jsCookie2.default.get('question_number') ? _jsCookie2.default.get('question_number') : 1;\n\nvar App = function (_Component) {\n  _inherits(App, _Component);\n\n  function App(props) {\n    _classCallCheck(this, App);\n\n    var _this = _possibleConstructorReturn(this, (App.__proto__ || Object.getPrototypeOf(App)).call(this, props));\n\n    _this.state = {\n      userName: user_name,\n      question: { question: '' },\n      questionNumber: 1,\n      answerNumber: 0,\n      isQuestion: true,\n      isAnswer: false,\n      userOptions: ''\n    };\n    return _this;\n  }\n\n  _createClass(App, [{\n    key: 'componentDidMount',\n    value: function componentDidMount() {\n      var _this2 = this;\n\n      this.fetchQuestion(question_number);\n      this.fetchUser();\n\n      socket.on('nextQuestion', function (obj) {\n        //WebSocketサーバーからnextQuestionを受け取った際の処理\n        console.log('obj=' + JSON.stringify(obj));\n        _jsCookie2.default.set('question_number', obj.question.id);\n        var question_number = _jsCookie2.default.get('question_number');\n        _this2.setState({\n          userName: _this2.state.userName,\n          question: obj,\n          questionNumber: question_number,\n          answerNumber: 0,\n          isQuestion: true,\n          isAnswer: false\n        });\n      });\n    }\n  }, {\n    key: 'startQuestion',\n    value: function startQuestion(name) {\n      var baseUrl = this.defaultBaseUrl();\n      axios.get(baseUrl + '/questions/1');\n    }\n  }, {\n    key: 'setStateUser',\n    value: function setStateUser(user_id) {\n      _jsCookie2.default.set('user_id', user_id);\n      _jsCookie2.default.set('user_name', JSON.parse(_jsCookie2.default.get('user_info'))[parseInt(user_id, 10) - 1].name);\n      var user_name = _jsCookie2.default.get('user_name');\n      this.setState({\n        userName: user_name\n      });\n    }\n  }, {\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var _this3 = this;\n\n      _question2.default.fetchQuestion(questionId).then(function (data) {\n        console.dir(data);\n        _this3.setState({\n          question: { question: data }\n        });\n      }).catch(function (error) {\n        alert('エラーが発生しました。');\n        console.log(error);\n      });\n    }\n  }, {\n    key: 'fetchUser',\n    value: function fetchUser() {\n      var _this4 = this;\n\n      _question2.default.fetchUser().then(function (data) {\n        console.dir(data);\n        _jsCookie2.default.set('user_info', data);\n        var userOptions = data.map(function (n) {\n          return _react2.default.createElement(\n            'option',\n            { key: n.id, value: n.id },\n            n.name\n          );\n        });\n        _this4.setState({\n          userOptions: userOptions\n        });\n      });\n    }\n  }, {\n    key: 'showAnswers',\n    value: function showAnswers(answers) {\n      var _this5 = this;\n\n      var answersComponents = [];\n      var alphabet = ['A', 'B', 'C', 'D'];\n\n      var _loop = function _loop(index, answer) {\n        answersComponents.push(_react2.default.createElement(\n          'li',\n          { key: answer.id, onClick: function onClick() {\n              return _this5.handleAnswerNumber(index + 1);\n            }, className: 'answer-list' },\n          _react2.default.createElement(\n            'span',\n            null,\n            alphabet[index],\n            ':'\n          ),\n          ' ',\n          answer.content\n        ));\n      };\n\n      var _iteratorNormalCompletion = true;\n      var _didIteratorError = false;\n      var _iteratorError = undefined;\n\n      try {\n        for (var _iterator = answers.entries()[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {\n          var _ref = _step.value;\n\n          var _ref2 = _slicedToArray(_ref, 2);\n\n          var index = _ref2[0];\n          var answer = _ref2[1];\n\n          _loop(index, answer);\n        }\n      } catch (err) {\n        _didIteratorError = true;\n        _iteratorError = err;\n      } finally {\n        try {\n          if (!_iteratorNormalCompletion && _iterator.return) {\n            _iterator.return();\n          }\n        } finally {\n          if (_didIteratorError) {\n            throw _iteratorError;\n          }\n        }\n      }\n\n      return answersComponents;\n    }\n  }, {\n    key: 'handleAnswerNumber',\n    value: function handleAnswerNumber(number) {\n      console.log(number);\n      this.setState({\n        answerNumber: number\n      });\n    }\n  }, {\n    key: 'answerQuestion',\n    value: function answerQuestion(answerNumber) {\n      var _this6 = this;\n\n      _question2.default.fetchAnswer(this.state.questionNumber, answerNumber, _jsCookie2.default.get('user_id')).then(function (data) {\n        _this6.setState({\n          isAnswer: true,\n          isQuestion: false\n        });\n      }).catch(function (error) {\n        alert('エラーが発生しました 。');\n        console.log(error);\n      });\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _this7 = this;\n\n      return _react2.default.createElement(\n        'div',\n        null,\n        _react2.default.createElement(\n          'div',\n          null,\n          _react2.default.createElement(\n            'div',\n            null,\n            this.state.userName != 'guest' && _react2.default.createElement(\n              'div',\n              null,\n              _react2.default.createElement(\n                'p',\n                { id: 'userName', className: 'text-center' },\n                this.state.userName,\n                '\\u306E\\u30C1\\u30E3\\u30EC\\u30F3\\u30B8'\n              ),\n              this.state.isQuestion && _react2.default.createElement(_Question2.default, {\n                question: this.state.question,\n                answerNumber: this.state.answerNumber,\n                showAnswers: function showAnswers(value) {\n                  return _this7.showAnswers(value);\n                },\n                answerQuestion: function answerQuestion(value) {\n                  return _this7.answerQuestion(value);\n                }\n              }),\n              this.state.isAnswer && _react2.default.createElement(\n                'div',\n                null,\n                this.state.questionNumber != 2 && _react2.default.createElement(\n                  'h1',\n                  { className: 'text-center' },\n                  '\\u6B21\\u306E\\u8CEA\\u554F\\u307E\\u3067\\u5C11\\u3005\\u304A\\u5F85\\u3061\\u304F\\u3060\\u3055\\u3044'\n                ),\n                this.state.questionNumber == 2 && _react2.default.createElement(\n                  'div',\n                  { className: 'text-center' },\n                  _react2.default.createElement(\n                    'h1',\n                    null,\n                    '\\u304A\\u75B2\\u308C\\u69D8\\u3067\\u3059\\uFF01\\u7D42\\u4E86\\u3067\\u3059\\uFF01'\n                  ),\n                  _react2.default.createElement(\n                    'p',\n                    null,\n                    '\\u7D50\\u679C\\u767A\\u8868\\u307E\\u3067\\u304A\\u5F85\\u3061\\u4E0B\\u3055\\u3044'\n                  )\n                )\n              )\n            ),\n            this.state.userName == 'guest' && _react2.default.createElement(\n              'div',\n              null,\n              _react2.default.createElement(\n                'div',\n                { className: 'question-content-area' },\n                '\\u3088\\u3046\\u3053\\u305D',\n                this.state.userName,\n                '\\u69D8\\u3001\\u4E0B\\u306E\\u9078\\u629E\\u80A2\\u304B\\u3089\\u81EA\\u5206\\u306E\\u540D\\u524D\\u3092\\u9078\\u3093\\u3067\\u304F\\u3060\\u3055\\u3044\\u3002'\n              ),\n              _react2.default.createElement(\n                'div',\n                { className: 'user-select-area' },\n                _react2.default.createElement(\n                  'select',\n                  { className: 'btn',\n                    onChange: function onChange(e) {\n                      return _this7.setStateUser(e.target.value);\n                    } },\n                  _react2.default.createElement(\n                    'option',\n                    { value: '\\u9078\\u629E\\u3057\\u3066\\u304F\\u3060\\u3055\\u3044', className: 'default-select' },\n                    '\\u9078\\u629E\\u3057\\u3066\\u304F\\u3060\\u3055\\u3044'\n                  ),\n                  this.state.userOptions\n                )\n              )\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return App;\n}(_react.Component);\n\nexports.default = App;\n\n//# sourceURL=webpack:///./src/components/App.js?");
+eval("\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _slicedToArray = function () { function sliceIterator(arr, i) { var _arr = []; var _n = true; var _d = false; var _e = undefined; try { for (var _i = arr[Symbol.iterator](), _s; !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"]) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; } return function (arr, i) { if (Array.isArray(arr)) { return arr; } else if (Symbol.iterator in Object(arr)) { return sliceIterator(arr, i); } else { throw new TypeError(\"Invalid attempt to destructure non-iterable instance\"); } }; }();\n\nvar _createClass = function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; }();\n\nvar _react = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _socket = __webpack_require__(/*! socket.io-client */ \"./node_modules/socket.io-client/lib/index.js\");\n\nvar _socket2 = _interopRequireDefault(_socket);\n\nvar _Question = __webpack_require__(/*! ./Question */ \"./src/components/Question.js\");\n\nvar _Question2 = _interopRequireDefault(_Question);\n\nvar _question = __webpack_require__(/*! ../api/question */ \"./src/api/question.js\");\n\nvar _question2 = _interopRequireDefault(_question);\n\nvar _user = __webpack_require__(/*! ../api/user */ \"./src/api/user.js\");\n\nvar _user2 = _interopRequireDefault(_user);\n\nvar _jsCookie = __webpack_require__(/*! js-cookie */ \"./node_modules/js-cookie/src/js.cookie.js\");\n\nvar _jsCookie2 = _interopRequireDefault(_jsCookie);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar socket = _socket2.default.connect('http://localhost:3005');\n\n_jsCookie2.default.remove('user_name');\n_jsCookie2.default.remove('question_number');\nvar user_name = _jsCookie2.default.get('user_name') ? _jsCookie2.default.get('user_name') : 'guest';\nvar user_id = _jsCookie2.default.get('user_id') ? _jsCookie2.default.get('user_id') : 0;\nvar question_number = _jsCookie2.default.get('question_number') ? _jsCookie2.default.get('question_number') : 1;\n\nvar App = function (_Component) {\n  _inherits(App, _Component);\n\n  function App(props) {\n    _classCallCheck(this, App);\n\n    var _this = _possibleConstructorReturn(this, (App.__proto__ || Object.getPrototypeOf(App)).call(this, props));\n\n    _this.state = {\n      userName: user_name,\n      question: { question: '' },\n      questionNumber: 1,\n      answerNumber: 0,\n      isQuestion: true,\n      isAnswer: false,\n      userOptions: ''\n    };\n    return _this;\n  }\n\n  _createClass(App, [{\n    key: 'componentDidMount',\n    value: function componentDidMount() {\n      var _this2 = this;\n\n      this.fetchQuestion(question_number);\n      this.fetchUser();\n\n      socket.on('nextQuestion', function (obj) {\n        //WebSocketサーバーからnextQuestionを受け取った際の処理\n        console.log('obj=' + JSON.stringify(obj));\n        _jsCookie2.default.set('question_number', obj.question.id);\n        var question_number = _jsCookie2.default.get('question_number');\n        _this2.setState({\n          userName: _this2.state.userName,\n          question: obj,\n          questionNumber: question_number,\n          answerNumber: 0,\n          isQuestion: true,\n          isAnswer: false\n        });\n      });\n    }\n  }, {\n    key: 'startQuestion',\n    value: function startQuestion(name) {\n      var baseUrl = this.defaultBaseUrl();\n      axios.get(baseUrl + '/questions/1');\n    }\n  }, {\n    key: 'setStateUser',\n    value: function setStateUser(user_id) {\n      _jsCookie2.default.set('user_id', user_id);\n      _jsCookie2.default.set('user_name', JSON.parse(_jsCookie2.default.get('user_info'))[parseInt(user_id, 10) - 1].name);\n      var user_name = _jsCookie2.default.get('user_name');\n      this.setState({\n        userName: user_name\n      });\n    }\n  }, {\n    key: 'fetchQuestion',\n    value: function fetchQuestion(questionId) {\n      var _this3 = this;\n\n      _question2.default.fetchQuestion(questionId).then(function (data) {\n        console.dir(data);\n        _this3.setState({\n          question: { question: data }\n        });\n      }).catch(function (error) {\n        alert('エラーが発生しました。');\n        console.log(error);\n      });\n    }\n  }, {\n    key: 'fetchUser',\n    value: function fetchUser() {\n      var _this4 = this;\n\n      _user2.default.fetchUser().then(function (data) {\n        console.dir(data);\n        _jsCookie2.default.set('user_info', data);\n        var userOptions = data.map(function (n) {\n          return _react2.default.createElement(\n            'option',\n            { key: n.id, value: n.id },\n            n.name\n          );\n        });\n        _this4.setState({\n          userOptions: userOptions\n        });\n      });\n    }\n  }, {\n    key: 'showAnswers',\n    value: function showAnswers(answers) {\n      var _this5 = this;\n\n      var answersComponents = [];\n      var alphabet = ['A', 'B', 'C', 'D'];\n\n      var _loop = function _loop(index, answer) {\n        answersComponents.push(_react2.default.createElement(\n          'li',\n          { key: answer.id, onClick: function onClick() {\n              return _this5.handleAnswerNumber(index + 1);\n            }, className: 'answer-list' },\n          _react2.default.createElement(\n            'span',\n            null,\n            alphabet[index],\n            ':'\n          ),\n          ' ',\n          answer.content\n        ));\n      };\n\n      var _iteratorNormalCompletion = true;\n      var _didIteratorError = false;\n      var _iteratorError = undefined;\n\n      try {\n        for (var _iterator = answers.entries()[Symbol.iterator](), _step; !(_iteratorNormalCompletion = (_step = _iterator.next()).done); _iteratorNormalCompletion = true) {\n          var _ref = _step.value;\n\n          var _ref2 = _slicedToArray(_ref, 2);\n\n          var index = _ref2[0];\n          var answer = _ref2[1];\n\n          _loop(index, answer);\n        }\n      } catch (err) {\n        _didIteratorError = true;\n        _iteratorError = err;\n      } finally {\n        try {\n          if (!_iteratorNormalCompletion && _iterator.return) {\n            _iterator.return();\n          }\n        } finally {\n          if (_didIteratorError) {\n            throw _iteratorError;\n          }\n        }\n      }\n\n      return answersComponents;\n    }\n  }, {\n    key: 'handleAnswerNumber',\n    value: function handleAnswerNumber(number) {\n      console.log(number);\n      this.setState({\n        answerNumber: number\n      });\n    }\n  }, {\n    key: 'answerQuestion',\n    value: function answerQuestion(answerNumber) {\n      var _this6 = this;\n\n      _question2.default.fetchAnswer(this.state.questionNumber, answerNumber, _jsCookie2.default.get('user_id')).then(function (data) {\n        _this6.setState({\n          isAnswer: true,\n          isQuestion: false\n        });\n      }).catch(function (error) {\n        alert('エラーが発生しました 。');\n        console.log(error);\n      });\n    }\n  }, {\n    key: 'render',\n    value: function render() {\n      var _this7 = this;\n\n      return _react2.default.createElement(\n        'div',\n        null,\n        _react2.default.createElement(\n          'div',\n          null,\n          _react2.default.createElement(\n            'div',\n            null,\n            this.state.userName != 'guest' && _react2.default.createElement(\n              'div',\n              null,\n              _react2.default.createElement(\n                'p',\n                { id: 'userName', className: 'text-center' },\n                this.state.userName,\n                '\\u306E\\u30C1\\u30E3\\u30EC\\u30F3\\u30B8'\n              ),\n              this.state.isQuestion && _react2.default.createElement(_Question2.default, {\n                question: this.state.question,\n                answerNumber: this.state.answerNumber,\n                showAnswers: function showAnswers(value) {\n                  return _this7.showAnswers(value);\n                },\n                answerQuestion: function answerQuestion(value) {\n                  return _this7.answerQuestion(value);\n                }\n              }),\n              this.state.isAnswer && _react2.default.createElement(\n                'div',\n                null,\n                this.state.questionNumber != 2 && _react2.default.createElement(\n                  'h1',\n                  { className: 'text-center' },\n                  '\\u6B21\\u306E\\u8CEA\\u554F\\u307E\\u3067\\u5C11\\u3005\\u304A\\u5F85\\u3061\\u304F\\u3060\\u3055\\u3044'\n                ),\n                this.state.questionNumber == 2 && _react2.default.createElement(\n                  'div',\n                  { className: 'text-center' },\n                  _react2.default.createElement(\n                    'h1',\n                    null,\n                    '\\u304A\\u75B2\\u308C\\u69D8\\u3067\\u3059\\uFF01\\u7D42\\u4E86\\u3067\\u3059\\uFF01'\n                  ),\n                  _react2.default.createElement(\n                    'p',\n                    null,\n                    '\\u7D50\\u679C\\u767A\\u8868\\u307E\\u3067\\u304A\\u5F85\\u3061\\u4E0B\\u3055\\u3044'\n                  )\n                )\n              )\n            ),\n            this.state.userName == 'guest' && _react2.default.createElement(\n              'div',\n              null,\n              _react2.default.createElement(\n                'div',\n                { className: 'question-content-area' },\n                '\\u3088\\u3046\\u3053\\u305D',\n                this.state.userName,\n                '\\u69D8\\u3001\\u4E0B\\u306E\\u9078\\u629E\\u80A2\\u304B\\u3089\\u81EA\\u5206\\u306E\\u540D\\u524D\\u3092\\u9078\\u3093\\u3067\\u304F\\u3060\\u3055\\u3044\\u3002'\n              ),\n              _react2.default.createElement(\n                'div',\n                { className: 'user-select-area' },\n                _react2.default.createElement(\n                  'select',\n                  { className: 'btn',\n                    onChange: function onChange(e) {\n                      return _this7.setStateUser(e.target.value);\n                    } },\n                  _react2.default.createElement(\n                    'option',\n                    { value: '\\u9078\\u629E\\u3057\\u3066\\u304F\\u3060\\u3055\\u3044', className: 'default-select' },\n                    '\\u9078\\u629E\\u3057\\u3066\\u304F\\u3060\\u3055\\u3044'\n                  ),\n                  this.state.userOptions\n                )\n              )\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return App;\n}(_react.Component);\n\nexports.default = App;\n\n//# sourceURL=webpack:///./src/components/App.js?");
 
 /***/ }),
